change data model FoodDto

+++ src/main/kotlin/com/example/model/food/FoodDto.kt
@@ -4,10 +4,10 @@ package com.example.model.food
 data class FoodDto(
     val image: String,
     var name: String,
-    val kiloCalories: Double,
-    val protein: Double,
-    val fat: Double,
-    val carbohydrate: Double,
+    val kiloCalories: Int,
+    val protein: Int,
+    val fat: Int,
+    val carbohydrate: Int,
     var count: Int = 0,
     var docId: String? = null
 ) {
